Severe libxml2 and xmlsec riot.
Always put <ds:Signature> as the last node since it looks like the right thing
to do (and even if the schema seems to say otherwise).

Try to clean xml messages and remove redundant xml namespace declarations.
Behind libxml2 back and cleaning memory properly.

We���re off the streets now / And back on the road / On the riot trail.