Be correct in dealing with varargs
Although in this cases it may be safe to omit va_end() that is not generally
so with an arbitrary compiler on an arbitrary platform.

Quoting from the spec: "Each invocation of va_start() must be matched by
a corresponding invocation of va_end() in the same function."
note the "must"

Checker tools like Coverity complain if va_start() is not always paired with
va_end(), so this patch mutes them.

License: MIT
Signed-off-by: Simo Sorce <simo@redhat.com>