Fixes for unchecked return values
Coverity complains that in a number of places errors are not checked.
Some of them are ok not to check so put a silencing (void).
Check errors that need to be checked.

Coverity also complains g_malloc() return is never checked but given it is
never checked anywahere let it be for now.

License: MIT
Signed-off-by: Simo Sorce <simo@redhat.com>