Fix LassoSoapHeader, was unable to serialize from XML.
The existing Lasso code never made use of SOAP headers because up
until now nothing used them. LassoSoapHeader was unable to serialize
from XML into a GList of LassoNode objects because it was missing one
of the necessary snippet flags. This corrects this omission and now
parsing a SOAP header will yield a sequence of LassoNode's.

Signed-off-by: John Dennis <jdennis@redhat.com>
License: MIT