LassoSamlp2IDPList is not list capable
LassoSamlp2IDPList is supposed to handle a list of LassoSamlp2IDPEntry
but in fact it had no list support. Change the snippet flag
SNIPPET_NODE to SNIPPET_LIST_NODES and add the special list comment on
the struct member so that the binding generator knows what type of
GList it is.

Signed-off-by: John Dennis <jdennis@redhat.com>
License: MIT