add inline implementation of lasso_log
lasso_log is a private function of lasso and as such cannot be
referenced by the loader.

This is equivalent to commit e0bda691 in the PHP binding which
exhibited the same problem.

lasso_log is referenced in jobject_to_gobject() because of
lasso_assign_gobject macro, which includes the lasso_release_gobject
macro which invokes the message macro which expands to lasso_log.

License: MIT
Signed-off-by: John Dennis <jdennis@redhat.com>